DSPBRIDGE: Change imode type to u8
imode is used to represent channel input or output mode.
It can take value either 0, 1 or 2. Hence changing it to u8.

+++ arch/arm/plat-omap/include/dspbridge/io_sm.h
@@ -123,7 +123,7 @@ void dsp_wdt_set_enable(bool);
 
 extern void io_request_chnl(struct io_mgr *hio_mgr,
 			    struct chnl_object *pchnl,
-			    u32 iMode, OUT u16 *pwMbVal);
+			    u8 iMode, OUT u16 *pwMbVal);
 
 /*
  *  ======== iosm_schedule ========
